admin: change old routes to new session routes
