+++ admin/app/components/sessions/list-items.hbs
@@ -30,7 +30,7 @@
       <tbody>
       {{#each @sessions as |session|}}
         <tr>
-          <td>{{session.id}}</td>
+          <td><LinkTo @route="authenticated.sessions.session" @model={{session.id}}> {{session.id}} </LinkTo></td>
           <td>{{session.certificationCenterName}}</td>
             {{#if (and session.certificationCenter session.certificationCenter.type)}}
               <td class="session-list__item--align-center">{{session.certificationCenter.type}}</td>
